sonarr_mgmt/requirements.txt
arrow==0.17.0
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
oauthlib==3.2.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
trakt.py==4.4.0
urllib3==2.2.1
